requirements.txt
pyTelegramBotAPI==3.7.3
python-dateutil==2.8.1
google-api-python-client==1.12.1
google-auth-oauthlib==0.4.1